backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
chromadb==0.5.23
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
PyPDF2==3.0.1
reportlab==4.2.5
Pillow==11.0.0
pydantic==2.10.3
python-dotenv==1.0.1
aiosqlite==0.20.0
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
